refactor: use db data and ssr for police_financial
